GUI improvements
The size of the bitmap on screen now matches its real size.

Automatically update the output on any change to the input.

Handle bad numbers.
The used to cause an exception.
Now it shows up as a pink background.

The font now knows the correct line width.
I had previously forgotten to use this input.

Implement the recommendedLineWidthButton.