Miscellaneous improvements.
Make the checkerboard look better.
It was fine for bright colors but failed for dark colors.
Those, especially pure black, displayed an all black background, no checkerboard.
I switched to a more appropriate color space.
That made the result more consistent.
And I made a bigger change.

Fix some warnings.
We we updating the URL too frequently.

Add an error message.
Display it when someone throws an Error.
Do not rethrow the Error.
That was causing some problems if there was an error processing the # on initialization.

Reorganize the code.
class Hash contains some items that were previously all separate variables.